refactor: split CLI entrypoint tests
